feat: successfully integrate favorite event feature

+++ app/src/main/java/com/raflis/final_submission_1_android_fundamental/ui/event_details/FavoriteEventAddDeleteViewModel.kt
@@ -1,24 +1,38 @@
 package com.raflis.final_submission_1_android_fundamental.ui.event_details
 
 import android.app.Application
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MediatorLiveData
 import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
 import com.raflis.final_submission_1_android_fundamental.data.local.entity.FavoriteEvent
 import com.raflis.final_submission_1_android_fundamental.data.local.repository.FavoriteEventRepository
+import kotlinx.coroutines.launch
 
 class FavoriteEventAddDeleteViewModel(application: Application) : ViewModel() {
     private val mFavoriteEventRepository: FavoriteEventRepository =
         FavoriteEventRepository(application)
 
-    fun insert(favoriteEvent: FavoriteEvent) {
-        mFavoriteEventRepository.insert(favoriteEvent)
+    private val _isFavorite = MediatorLiveData<Boolean>()
+
+    fun isFavoriteEvent(eventId: Int?): LiveData<Boolean> {
+        val favoriteEventLiveData = mFavoriteEventRepository.getFavoriteEventById(eventId)
+        _isFavorite.addSource(favoriteEventLiveData) { favoriteEvent ->
+            _isFavorite.value = favoriteEvent != null
+        }
+        return _isFavorite
     }
 
-    fun update(favoriteEvent: FavoriteEvent) {
-        mFavoriteEventRepository.update(favoriteEvent)
-    }
-
-    fun delete(favoriteEvent: FavoriteEvent) {
-        mFavoriteEventRepository.delete(favoriteEvent)
+    fun toggleFavoriteEvent(favoriteEvent: FavoriteEvent) {
+        viewModelScope.launch {
+            val isFavorite = _isFavorite.value ?: false
+            if (isFavorite) {
+                mFavoriteEventRepository.delete(favoriteEvent)
+            } else {
+                mFavoriteEventRepository.insert(favoriteEvent)
+            }
+            _isFavorite.value = !isFavorite
+        }
     }
 
 }
